requirements.txt
addict==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
babel==2.18.0
blinker==1.9.0
blis==1.3.3
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
csvw==3.7.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.13
defusedxml==0.7.1
dlinfo==2.0.0
docopt==0.6.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
espeakng-loader==0.2.4
fastapi==0.124.4
filelock==3.24.2
Flask==3.1.2
flask-cors==6.0.1
fsspec==2026.2.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
hyperframe==6.1.0
idna==3.11
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kokoro==0.9.4
language-tags==1.2.0
llvmlite==0.46.0
loguru==0.7.3
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
misaki==0.9.4
more-itertools==10.8.0
mpmath==1.3.0
murmurhash==1.0.15
networkx==3.6.1
num2words==0.5.14
numba==0.63.1
numpy==2.3.5
odfpy==1.4.1
openai-whisper==20250625
packaging==26.0
phonemizer-fork==3.3.2
pillow==12.1.0
preshed==3.0.12
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
PyYAML==6.0.3
rdflib==7.6.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rfc3986==1.5.0
rich==14.3.2
rpds-py==0.30.0
safetensors==0.7.0
segments==2.3.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
soundfile==0.13.1
spacy==3.8.11
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
starlette==0.50.0
sympy==1.14.0
termcolor==3.3.0
thinc==8.3.10
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.10.0+cu128
torchaudio==2.10.0+cu128
torchvision==0.25.0+cu128
tqdm==4.67.3
transformers==5.2.0
typer==0.24.0
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.0
uvicorn==0.38.0
wasabi==1.1.3
weasel==0.4.3
Werkzeug==3.1.4
win32_setctime==1.2.0
wrapt==2.1.1
